+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Fixed
+ - Fix Steam library folder discovery by using case-insensitive directory name matching
+
 ## [1.13.0] - 2025-08-11
 ### Added
  - Improve compatibility by setting additional Proton related environment variables when applicable